+++ changelog.md
@@ -42,6 +42,7 @@
 * **[UX]** Default settings are now loaded from Default.zip
 * **[Autoplanner]** Plan Air-to-Air Escorts for AWACS & Tankers
 * **[Package Planning]** Ability to plan recovery tanker flights
+* **[Modding]** Support for Bandit's cloud presets mod (v15)
 
 ## Fixes
 * **[UI/UX]** A-10A flights can be edited again